Add Files for upload from Today's Meeting
Ali and Juandalyn met today and selected cities to use for the project which are in the ABJB_localchikdata.csv. We found that the temperature mean values (MNMT) were in tenths of degree Celsius (via the NOAAabbreviations.pdf doc).  We converted the values in the CharlestonSC_climatedata_example.xlsx file.  

We think that the temp data may be in tenths degree Celsius because it may help to distinguish temperature ranges.  We also are thinking about grouping probabilities of transmission by low (<=0.24, medium (0.25-0.5), and high (>=0.5)